Refactor Athena class and menthods into one shared/utils file
